Use log4net instead of custom logging system.
Rewrote the QuadTree system; it uses the same core algorithms
but is now significantly more efficient. However, it assumes
all tracked items do not have a dimension and are instead just
point items.

Introduced WasSelected/WasDeselected events and added an event
responder that grows/shrinks the node on said events.