Remove NetSerializer.
Even though it is faster than Neon.Serialization, it uses code
generation for its speed. This means that it is not usable on iOS,
and given the amount of serialization that Neon's networking model
requires (very little), the additional overhead of
Neon.Serialization should not prove to be an issue.