IGameEngine uses Tasks instead of WaitHandles
This could be a problem for cross AppDomain communication, but
cross AppDomain communication still has to deal with the overhead
of all the event serialization, which is currently deemed too high.

This change has one big benefit: if the running task has an exception
(such as system throwing), then the main thread will actually see
the exception, instead of it being silently ignored.