Introduce non-versioned Data
These are data types where there is only one state. The previous
value for that state cannot be viewed and the value of the
data cannot be viewed after it has been modified.

Non-versioned data seems to go against the core philosophy of
Forge.Entities, but it is extremely useful for certain data types
which are fundamentally not versioned, ie, data that will only be
used in one system. It better captures the idea that the data type
cannot be backwards queried instead of having the client code
silently express that idea; nonversioned data makes that explicit.