Contains=true and Current=value for removed data
This fixes a problematic scenario with non-versioned data. If you
remove the non-versioned data, in the next update when you're
trying to cleanup Current(accessor) will throw a NoSuchData
exception and Previous(accessor) will throw a
PreviousRequiresVersionedData exception. This change allows for
Current to require the last known value for the data.