Add Trigger.OnEngineLoaded
This allows for arbitrary code execution when starting a game engine. The
running code is not allowed to modify game state, however, as it is
executed only when a game engine has been initialized (for example, when a
new game has started or when a replay has been loaded or when a save file
has been loaded).